+++ js/src/main/scala/spatutorial/client/SPAMain.scala
@@ -4,7 +4,7 @@ import japgolly.scalajs.react.React
 import japgolly.scalajs.react.extra.router.BaseUrl
 import org.scalajs.dom
 import spatutorial.client.modules.MainRouter
-
+import spatutorial.client.logger._
 import scala.scalajs.js
 import scala.scalajs.js.annotation.JSExport
 
@@ -12,6 +12,11 @@ import scala.scalajs.js.annotation.JSExport
 object SPAMain extends js.JSApp {
   @JSExport
   def main(): Unit = {
+    log.warn("Application starting")
+    // send log messages also to the server
+    log.enableServerLogging("/logging")
+    log.info("This message goes to server as well")
+
     // build a baseUrl, this method works for both local and server addresses (assuming you use #)
     val baseUrl = BaseUrl(dom.window.location.href.takeWhile(_ != '#'))
     val router = MainRouter.router(baseUrl)