+++ js/src/main/scala/spatutorial/client/logger/package.scala
@@ -1,0 +1,7 @@
+package spatutorial.client
+
+package object logger {
+  private val defaultLogger = LoggerFactory.getLogger("Log")
+
+  def log = defaultLogger
+}